Avoid duplicating NewDatabaseForTest
